Fix danger duration minutes

+++ src/main/java/com/example/flood_alert/dbo/response/AreaRiskSnapshotResponse.java
@@ -18,7 +18,7 @@ public class AreaRiskSnapshotResponse {
     Double iotRiskScore;
     Double predictionProbability;
     Double dangerRatio;
-    Integer dangerDurationMinutes;
+    BigDecimal dangerDurationMinutes;
     BigDecimal waterRiseRatePerMinute;
     Integer dangerAggregateCount;
     Double dangerPercent;
